requirements.txt
streamlit
python-dotenv
# LlamaIndex core and integrations
llama-index>=0.10.30
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-chroma
llama-index-readers-file
# Vector DB
chromadb
# PDF parsing for SimpleDirectoryReader
pypdf
# Optional: for local embeddings (uncomment if you want to use them)
# sentence-transformers
# llama-index-embeddings-huggingface